Performance optimization

+++ src/helper.hpp
@@ -6,7 +6,7 @@ void changePixel(Screen& screen, Pos2 pos, uint32_t color);
 
 void drawLine(Screen& screen, Pos2 P0, Pos2 P1, uint32_t color=WHITE);
 
-std::vector<int> interpolatePoints(int i0, int o0, int i1, int o1);
+std::vector<short int> interpolatePoints(short int i0, short int o0, short int i1, short int o1);
 
 void drawWireframeTriangle(Screen& screen, Pos2 p0, Pos2 p1, Pos2 p2);
